Fix two CatFIM bugs
This fixes a bug in the get_metadata() function in /tools/tools_shared_functions.py that arose because of a WRDS update. Previously the metadata_source response was returned as independent variables, but now it is are returned a list of strings. Another issue was observed where the EVALUATED_SITES_CSV variable was being misdefined (at least on the development VM) through the OS environmental variable setting.

- Changed parsing of WRDS metadata_sources to account for new list type.
- Changed the way the EVALUATED_SITES_CSVpath is defined from OS environmental setting to a relative path that will work within Docker container.

This resolves #445.