Rename BARC variables
Renames the BARC modified variables that are exported to src_full_crosswalked.csv to replace the original variables. The default/original variables are renamed with orig_ prefix. This change is needed to ensure downstream uses of the src_full_crosswalked.csv are able to reference the authoritative version of the channel geometry variables (i.e. BARC-adjust where available).

- In src_full_crosswalked.csv, default/original variables are renamed with orig_ prefix and SA_div is renamed to SA_div_flag.

This resolves #418.